chore: consolidate components
