+++ app/components/scroll/ScrollMintPage.tsx
@@ -1,20 +1,54 @@
 import { useContext, useEffect, useMemo, useState } from "react";
 import { useMessage } from "../../hooks/useMessage";
-import { useWalletStore } from "../../context/walletStore";
 import { CeramicContext } from "../../context/ceramicContext";
 import { useAttestation } from "../../hooks/useAttestation";
-import { EasPayload, Passport, Stamp } from "@gitcoin/passport-types";
+import { PROVIDER_ID, Passport, Stamp } from "@gitcoin/passport-types";
 import {
+  badgeContractInfo,
   scrollCampaignBadgeProviderInfo,
   scrollCampaignBadgeProviders,
   scrollCampaignChain,
 } from "../../config/scroll_campaign";
-import { ProviderWithTitle, getEarnedBadges } from "../ScrollCampaign";
-import { iamUrl } from "../../config/stamp_config";
-import { jsonRequest } from "../../utils/AttestationProvider";
+import { ProviderWithTitle } from "../ScrollCampaign";
 import { ScrollCampaignPage } from "./ScrollCampaignPage";
 import { LoadingBarSection, LoadingBarSectionProps } from "../LoadingBar";
 import { LoadButton } from "../LoadButton";
+
+export const getEarnedBadges = (badgeStamps: Stamp[]): ProviderWithTitle[] => {
+  if (badgeStamps.length === 0) {
+    return [];
+  }
+  return badgeContractInfo.map((contract) => {
+    const relevantStamps = badgeStamps.filter((stamp) =>
+      contract.providers.some(({ name }) => name === stamp.provider)
+    );
+
+    if (relevantStamps.length === 0) {
+      return {
+        title: contract.title,
+        name: "No Provider" as PROVIDER_ID,
+        image: "",
+        level: 0,
+      };
+    }
+
+    const highestLevelProvider = relevantStamps.reduce(
+      (highest, stamp) => {
+        const provider = contract.providers.find(({ name }) => name === stamp.provider);
+        if (provider && provider.level > highest.level) {
+          return provider;
+        }
+        return highest;
+      },
+      { level: -1, name: "No Provider" as PROVIDER_ID, image: "" }
+    );
+
+    return {
+      title: contract.title,
+      ...highestLevelProvider,
+    };
+  });
+};
 
 export const ScrollMintBadge = ({
   onMintBadge,
@@ -23,9 +57,7 @@ export const ScrollMintBadge = ({
 }) => {
   const { failure } = useMessage();
   const { database } = useContext(CeramicContext);
-  const address = useWalletStore((state) => state.address);
-  const { getNonce, issueAttestation, needToSwitchChain } = useAttestation({ chain: scrollCampaignChain });
-  const [syncingToChain, setSyncingToChain] = useState(false);
+  const { needToSwitchChain } = useAttestation({ chain: scrollCampaignChain });
 
   const [passport, setPassport] = useState<Passport | undefined>(undefined);
 
@@ -43,7 +75,7 @@ export const ScrollMintBadge = ({
         }
       }
     })();
-  }, [database]);
+  }, [database, failure]);
 
   const badgeStamps = useMemo(
     () => (passport ? passport.stamps.filter(({ provider }) => scrollCampaignBadgeProviders.includes(provider)) : []),
@@ -74,7 +106,7 @@ export const ScrollMintBadge = ({
           {} as Record<string, { level: number; credential: Stamp }>
         )
       ).map(({ credential }) => credential),
-    [badgeStamps, deduplicatedBadgeStamps]
+    [deduplicatedBadgeStamps]
   );
 
   const earnedBadges = getEarnedBadges(badgeStamps);
@@ -115,12 +147,10 @@ export const ScrollMintBadge = ({
           <LoadButton
             variant="custom"
             onClick={() => onMintBadge(earnedBadges)}
-            isLoading={loading || syncingToChain}
+            isLoading={loading}
             className="text-color-1 text-lg font-bold bg-[#FF684B] hover:brightness-150 py-3 transition-all duration-200"
           >
-            <div className="flex flex-col items-center justify-center">
-              {syncingToChain ? "Minting..." : "Mint Badge"}
-            </div>
+            <div className="flex flex-col items-center justify-center">{"Mint Badge"}</div>
           </LoadButton>
           {needToSwitchChain && (
             <div className="text-[#FF684B] mt-4">
